Added machine learning classes.
Due to issues with implementing the C++ code and issue with not being
able to use classes, given that it's only one week left, I have now
spent the night re-writing most of the C/C++ code into Java code, which
hopefully will execute flawlessly. The only remaining part is to create
a more proper NeuralNetwork and SelfOrganizingMap class code.